# web/public/api-docs/openapi.yaml
openapi: 3.1.0
info:
  title: aladdin API
  version: v1
  description: |-
    由 `make api-docs` 从 proto 生成，请勿手改。

    每个方法的鉴权要求写在它说明的开头；同一份信息也以 `x-aladdin-auth` 扩展给出，供机器读取。
paths:
...
  /aladdin.identity.v1.IdentityService/Refresh:
    post:
      x-aladdin-auth:
        kind: public
      tags:
        - aladdin.identity.v1.IdentityService
      summary: Refresh
      description: |-
        **鉴权**：公开接口，无需认证。

        刷新访问凭证。
      operationId: aladdin.identity.v1.IdentityService.Refresh
      parameters:
        - name: Connect-Protocol-Version
          in: header
          required: true
          schema:
            $ref: '#/components/schemas/connect-protocol-version'
        - name: Connect-Timeout-Ms
          in: header
          schema:
            $ref: '#/components/schemas/connect-timeout-header'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/aladdin.identity.v1.RefreshRequest'
        required: true
      responses:
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/connect.error'
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/aladdin.identity.v1.RefreshResponse'
...
  /aladdin.identity.v1.IdentityService/GetSessionPermissions:
    post:
      x-aladdin-auth:
        kind: authenticated_only
      tags:
        - aladdin.identity.v1.IdentityService
      summary: GetSessionPermissions
      description: |-
        **鉴权**：需要登录。任何已认证主体均可调用。

        返回当前主体在其可及作用域下展开后的权限码集合。
         这是前端会话权限的唯一来源：前端拿到的是已展开的最终集合，
         因此前端不需要（也不允许）自行实现继承、通配与作用域包含逻辑。
      operationId: aladdin.identity.v1.IdentityService.GetSessionPermissions
      parameters:
        - name: Connect-Protocol-Version
          in: header
          required: true
          schema:
            $ref: '#/components/schemas/connect-protocol-version'
        - name: Connect-Timeout-Ms
          in: header
          schema:
            $ref: '#/components/schemas/connect-timeout-header'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/aladdin.identity.v1.GetSessionPermissionsRequest'
        required: true
      responses:
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/connect.error'
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/aladdin.identity.v1.GetSessionPermissionsResponse'
...
components:
  schemas:
...
    connect-protocol-version:
      type: number
      title: Connect-Protocol-Version
      enum:
        - 1
      description: Define the version of the Connect protocol
      const: 1
    connect-timeout-header:
      type: number
      title: Connect-Timeout-Ms
      description: Define the timeout, in ms
    connect.error:
      type: object
      properties:
        code:
          type: string
          examples:
            - CodeNotFound
          enum:
            - CodeCanceled
            - CodeUnknown
            - CodeInvalidArgument
            - CodeDeadlineExceeded
            - CodeNotFound
            - CodeAlreadyExists
            - CodePermissionDenied
            - CodeResourceExhausted
            - CodeFailedPrecondition
            - CodeAborted
            - CodeOutOfRange
            - CodeInternal
            - CodeUnavailable
            - CodeDataLoss
            - CodeUnauthenticated
          description: The status code, which should be an enum value of [google.rpc.Code][google.rpc.Code].
        message:
          type: string
          description: A developer-facing error message, which should be in English. Any user-facing error message should be localized and sent in the [google.rpc.Status.details][google.rpc.Status.details] field, or localized by the client.
        detail:
          $ref: '#/components/schemas/google.protobuf.Any'
      title: Connect Error
      additionalProperties: true
      description: 'Error type returned by Connect: https://connectrpc.com/docs/go/errors/#http-representation'
    google.protobuf.Any:
      type: object
      properties:
        type:
          type: string
        value:
          type: string
          format: binary
        debug:
          type: object
          additionalProperties: true
      additionalProperties: true
      description: Contains an arbitrary serialized message along with a @type that describes the type of the serialized message.
...
    aladdin.identity.v1.GetSessionPermissionsRequest:
      type: object
      properties:
        scope:
          type: string
          title: scope
          description: 期望的作用域。留空时使用凭证的默认作用域。
      title: GetSessionPermissionsRequest
      additionalProperties: false
    aladdin.identity.v1.GetSessionPermissionsResponse:
      type: object
      properties:
        scope:
          type: string
          title: scope
        permissions:
          type: array
          items:
            type: string
          title: permissions
          description: 已展开的最终权限码集合。
      title: GetSessionPermissionsResponse
      additionalProperties: false
...
    aladdin.identity.v1.RefreshRequest:
      type: object
      properties:
        accessToken:
          type: string
          title: access_token
      title: RefreshRequest
      additionalProperties: false
    aladdin.identity.v1.RefreshResponse:
      type: object
      properties:
        accessToken:
          type: string
          title: access_token
        expiresAt:
          type: string
          title: expires_at
      title: RefreshResponse
      additionalProperties: false
...